# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ouvrier-Express",
        "description": "Application de gestion de relation prestataire-client",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ajouterCategorie": {
            "post": {
                "tags": [
                    "Catégorie de service"
                ],
                "summary": "Ajouter une catégorie de service",
                "operationId": "b19e4cac34b494bfb9d92f020592cc3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelleCategorie": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Fichier de photo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie de service ajoutée avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/affichCategorie/{id}": {
            "get": {
                "tags": [
                    "Catégorie de service"
                ],
                "summary": "Details d'une categorie de service",
                "operationId": "6c1af176b8e393b3eed687d026d4e4e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/modifCategorie/{id}": {
            "patch": {
                "tags": [
                    "Catégorie de service"
                ],
                "summary": "Modificier une catégorie de service",
                "operationId": "8a7d376c403f297988452c39d9b20d3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelleCategorie": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Fichier de photo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie de service modifiée avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supprimCategorie/{id}": {
            "patch": {
                "tags": [
                    "Catégorie de service"
                ],
                "summary": "Supprimer une catégorie de service",
                "operationId": "31097b1b5915aabb3d65d75847cd1d09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Suppression d'une catégorie à partir de l'id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                }
            }
        },
...
        "/api/ajoutComment": {
            "post": {
                "tags": [
                    "Commentaire sur un prestataire"
                ],
                "summary": "Ajouter un commentaire sur un prestataire",
                "operationId": "2ed4cc14cf97335461f00c77388cd6d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "statut_evaluation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire ajoutée avec succés"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/aaffichComment/{id}": {
            "get": {
                "tags": [
                    "Commentaire sur un prestataire"
                ],
                "summary": "Afficher un commentaire",
                "operationId": "aeb3ae0e87d0b0ef9b267db2f57e110f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id du commentaire à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/modifCategorie/{commentaire}": {
            "patch": {
                "tags": [
                    "Commentaire sur un prestataire"
                ],
                "summary": "Modificier un commentaire",
                "operationId": "883d4a74037fb67eb42243cb32036b70",
                "parameters": [
                    {
                        "name": "commentaire",
                        "in": "path",
                        "description": "ID du commentaire à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "statut_evaluation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire modifié avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supprimComment/{id}": {
            "patch": {
                "tags": [
                    "Commentaire sur un prestataire"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "d0a58e2e7dc4bb558909ccd0bda98434",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Supprimer un commentaire à partir de l'id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}